pyproject.toml
[project]
name = "nova-agent"
version = "0.1.0"
description = "Local Autonomus Agent"
authors = [
    {name = "Julian Padron",email = "julian.padronnunez03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "typer[all] (>=0.24.1,<0.25.0)",
    "rich (>=14.3.3,<15.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0) ; python_version >= \"3.11\" and python_full_version < \"4.0.0\"",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "google-search-results (>=2.4.2,<3.0.0)",
    "textual (>=8.1.1,<9.0.0)",
    "psutil (>=7.2.2,<8.0.0)"
]

[project.scripts]
nova = "nova_agent.ui:run"

[tool.poetry]
# ESTO ES LO QUE ARREGLA EL ERROR DE MÓDULO:
# Le dice a Poetry que el código de 'nova_agent' vive dentro de 'src'
packages = [{include = "nova_agent", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"